lib/{uksglist,uknetdev}: Move `uk_sglist_append_netbuf` to `uknetdev`
Rename `uk_sglist_append_netbuf` to `uk_netbuf_sglist_append`
and move it to the libuknetdev, since that is its only client.
